move `indicatif` to only live in `cli`
