+++ cli/src/main.rs
@@ -2,9 +2,10 @@ mod args;
 mod execve;
 mod help;
 
-use std::{error::Error, time::Duration};
+use std::{error::Error, fmt::Write, sync::Arc, time::Duration};
 
 use execve::execve;
+use indicatif::{ProgressBar, ProgressState, ProgressStyle};
 use libpkgx::{
     config::Config, env, hydrate::hydrate, install_multi, pantry_db, resolve::resolve, sync,
     types::PackageReq, utils,
@@ -149,7 +150,11 @@ async fn main() -> Result<(), Box<dyn Error>> {
 
     let mut installations = resolution.installed;
     if !resolution.pending.is_empty() {
-        let installed = install_multi::install_multi(&resolution.pending, &config, spinner).await?;
+        let pb = spinner.map(|spinner| {
+            configure_bar(&spinner);
+            Arc::new(MultiProgressBar { pb: spinner })
+        });
+        let installed = install_multi::install_multi(&resolution.pending, &config, pb).await?;
         installations.extend(installed);
     }
 
@@ -250,5 +255,85 @@ async fn which(cmd: &String, conn: &Connection) -> Result<String, WhichError> {
         return Err(WhichError::CmdNotFound(cmd.clone()));
     } else {
         return Err(WhichError::MultipleProjects(cmd.clone(), candidates));
+    }
+}
+
+struct MultiProgressBar {
+    pb: ProgressBar,
+}
+
+impl libpkgx::install_multi::ProgressBarExt for MultiProgressBar {
+    fn inc(&self, n: u64) {
+        self.pb.inc(n);
+    }
+
+    fn inc_length(&self, n: u64) {
+        self.pb.inc_length(n);
+    }
+}
+
+// ProgressBar is Send + Sync
+unsafe impl Send for MultiProgressBar {}
+unsafe impl Sync for MultiProgressBar {}
+
+fn configure_bar(pb: &ProgressBar) {
+    pb.set_length(1);
+    pb.set_style(
+        ProgressStyle::with_template(
+            "{elapsed:.dim} ❲{wide_bar:.red}❳ {percent}% {bytes_per_sec:.dim} {bytes:.dim}",
+        )
+        .unwrap()
+        .with_key("elapsed", |state: &ProgressState, w: &mut dyn Write| {
+            let s = state.elapsed().as_secs_f64();
+            let precision = precision(s);
+            write!(w, "{:.precision$}s", s, precision = precision).unwrap()
+        })
+        .with_key("bytes", |state: &ProgressState, w: &mut dyn Write| {
+            let (right, divisor) = pretty_size(state.len().unwrap());
+            let left = state.pos() as f64 / divisor as f64;
+            let leftprecision = precision(left);
+            write!(
+                w,
+                "{:.precision$}/{}",
+                left,
+                right,
+                precision = leftprecision
+            )
+            .unwrap()
+        })
+        .progress_chars("⚯ "),
+    );
+    pb.enable_steady_tick(Duration::from_millis(50));
+}
+
+fn pretty_size(n: u64) -> (String, u64) {
+    let units = ["B", "KiB", "MiB", "GiB", "TiB", "PiB", "EiB", "ZiB", "YiB"];
+    let mut size = n as f64;
+    let mut i = 0;
+    let mut divisor = 1;
+
+    while size > 1024.0 && i < units.len() - 1 {
+        size /= 1024.0;
+        i += 1;
+        divisor *= 1024;
+    }
+
+    let formatted = format!(
+        "{:.precision$} {}",
+        size,
+        units[i],
+        precision = precision(size)
+    );
+
+    (formatted, divisor)
+}
+
+fn precision(n: f64) -> usize {
+    if n < 10.0 {
+        2
+    } else if n < 100.0 {
+        1
+    } else {
+        0
     }
 }
