+++ lib/Cargo.toml
@@ -24,7 +24,6 @@ async-compression = { version = "0.4", features = ["tokio", "gzip", "xz"] }
 tokio-tar = "0.3.1"
 tokio-util = { version = "0.7.13", features = ["compat"] }
 futures = "0.3.31"
-indicatif = "0.17.9"
 lazy_static = "1.5.0"
 nix = { version = "0.29.0", features = ["process"] }
 fs2 = "0.4.3"
